spring-homework-05/src/main/java/ru/otus/homework/repository/genre/GenreRepositoryJpaImpl.java
package ru.otus.homework.repository.genre;

import org.springframework.stereotype.Repository;
import ru.otus.homework.domain.Genre;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.persistence.Query;
import javax.persistence.TypedQuery;
import java.util.List;
import java.util.Optional;

@Repository
public class GenreRepositoryJpaImpl implements GenreRepositoryJpa {

    @PersistenceContext
    private EntityManager em;


    @Override
    public Genre insert(Genre genre) {
       if(genre.getId()==0){
           em.persist(genre);
           em.flush();
           return genre;
       }else {
           return em.merge(genre);
       }
    }

    @Override
    public Optional<Genre> findById(long id) {
        return Optional.ofNullable(em.find(Genre.class,id));
    }

    @Override
    public List<Genre> findAll() {
        TypedQuery<Genre> query = em.createQuery("select g from Genre g",Genre.class);
        return query.getResultList();
    }

    @Override
    public void deleteById(long id) {
        Query query = em.createQuery("delete from Genre g where g.id = :id");
        query.setParameter("id",id);
        query.executeUpdate();
    }
}
